requirements.txt
Flask==2.3.3
openai==1.30.1
pinecone-client==3.2.2
python-dotenv==1.0.1
tqdm==4.66.4
